frontend/requirements.txt
#!/bin/bash

# Streamlit requirements for deployment
streamlit
requests
python-dotenv
pandas
